+++ uttt-parent/core/src/main/java/com/uttt/core/game/prebuilt/PlayerPredictable.java
@@ -1,4 +1,4 @@
-package com.uttt.core.game;
+package com.uttt.core.game.prebuilt;
 
 import com.uttt.common.RepeatableRandom;
 import com.uttt.core.board.Token;
