+++ uttt-parent/core/src/main/java/com/uttt/core/game/prebuilt/PlayerUnpredictable.java
@@ -1,4 +1,4 @@
-package com.uttt.core.game;
+package com.uttt.core.game.prebuilt;
 
 import java.util.Random;
