+++ uttt-parent/core/src/main/java/com/uttt/core/game/prebuilt/PlayerWeight.java
@@ -1,4 +1,4 @@
-package com.uttt.core.game;
+package com.uttt.core.game.prebuilt;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -14,6 +14,7 @@ import com.uttt.core.board.Board;
 import com.uttt.core.board.Position;
 import com.uttt.core.board.Token;
 import com.uttt.core.board.Node.Status;
+import com.uttt.core.game.Move;
 import com.uttt.core.player.Player;
 
 public class PlayerWeight extends Player {
@@ -50,7 +51,7 @@ public class PlayerWeight extends Player {
 			Move subMove = weightedMove(pfx + "|", log, constraint.derefBoard(), null);
 
 			for (Position cursorConstraint = constraint; cursorConstraint != null; cursorConstraint = cursorConstraint.getBoard().getPosition()) {
-				subMove = new Move(cursorConstraint.getRow(), cursorConstraint.getCol(), subMove);
+                subMove = subMove.within(cursorConstraint.getRow(), cursorConstraint.getCol());
 			}
 			myMove = subMove;
 		} else {
@@ -80,13 +81,11 @@ public class PlayerWeight extends Player {
 					+ "myPlayPos=" + myPlayPos + "; " //
 			);
 
-			final Move subMove;
 			if (subBoard.isBottom()) {
-				subMove = null;
+                myMove = Move.newMove(myPlayPos.getRow(), myPlayPos.getCol());
 			} else {
-				subMove = weightedMove(pfx + "|", log, myPlayPos.derefBoard(), null);
+                myMove = weightedMove(pfx + "|", log, myPlayPos.derefBoard(), null).within(myPlayPos.getRow(), myPlayPos.getCol());
 			}
-			myMove = new Move(myPlayPos.getRow(), myPlayPos.getCol(), subMove);
 		}
 		log.trace(pfx //
 				+ "myMove=" + myMove + "; " //
