+++ uttt-parent/core/src/test/java/com/uttt/core/game/GameTest.java
@@ -8,7 +8,7 @@ import org.junit.Test;
 import com.uttt.common.UtttException;
 import com.uttt.core.board.Token;
 import com.uttt.core.game.Game;
-import com.uttt.core.game.PlayerPredictable;
+import com.uttt.core.game.prebuilt.PlayerPredictable;
 import com.uttt.core.player.Player;
 
 public class GameTest {
