+++ uttt-parent/core/src/test/java/com/uttt/core/game/WeightTest.java
@@ -8,7 +8,7 @@ import com.uttt.common.StackFrameUtil;
 import com.uttt.core.board.Board;
 import com.uttt.core.board.Token;
 import com.uttt.core.game.Move;
-import com.uttt.core.game.PlayerWeight;
+import com.uttt.core.game.prebuilt.PlayerWeight;
 import com.uttt.core.player.Player;
 
 public class WeightTest {
